Add test_side_by_side.py
